// smart_sale/types/bankAccount/BankAccount.ts
export interface BankAccount {

    ACCOUNTNO: string;
    ACCOUNTTYPE: string;
    ACHOLDERNAME: string;
    ADDRESS: string;
    BANKAC: string;
    BANKNAME: string;
    BRANCHNAME: string;
    OPENINGBALANCE: string;
    REMARKS: string;
}

export interface BankAccountMast {
    ENTRYNO:number,
    ACCOUNTNO:string;
    ACCOUNTTYPE:string;
    ACHOLDERNAME: string;
    ADDRESS: string;
    BANKAC:string;
    BANKNAME:string;
    BRANCHNAME:string;
    OPENINGBALANCE:string;
    REMARKS:string;
}